deleted AppLogForm, changed the AppLogTable to AppLogHistory, properly fixed delete and post http request without refreshing page.
