+++ client/src/components/AppLogHistory/AppLogHistory.js
@@ -1,9 +1,9 @@
 import React, { Component } from 'react';
-import {Table, TableBody,TableRow, TableHeader, TableHeaderColumn,TableFooter} from 'material-ui/Table';
+import {Table, TableBody,TableRow, TableHeader, TableHeaderColumn} from 'material-ui/Table';
 import FlatButton from 'material-ui/FlatButton';
+import TextField from 'material-ui/TextField';
 import axios from 'axios';
 import UserProfile from '../UserProfile/UserProfile';
-import AppLogForm from '../AppLogForm/AppLogForm';
 /*Need to clean this code before April begins */
 class AppLogTable extends Component{
     constructor(props){
@@ -11,12 +11,26 @@ class AppLogTable extends Component{
         this.state={
             applications:[],
             //this is for the material-ui
+            
             displaySelectAll: false,
             adjustForCheckbox: true,
-            selected:''
+            selected:'',
+            
+            //states are for post
+            company: '',
+            role:'',
+            status:''
         };
        this.handleRowSelection=this.handleRowSelection.bind(this); 
        this.handleDelete=this.handleDelete.bind(this);
+
+
+       //states for post and update
+       this.handleCompany=this.handleCompany.bind(this);
+       this.handleRole=this.handleRole.bind(this);
+       this.handleStatus=this.handleStatus.bind(this);
+       this.handleSubmit=this.handleSubmit.bind(this);
+       //this.handleModify=this.handleModify.bind(this);
     };
     componentDidMount(){
         const url='/applications/';
@@ -26,15 +40,11 @@ class AppLogTable extends Component{
         });
     }
     handleRowSelection(key){
-        console.log('why you here.');
-        let currentKey = this.state.applications[key]._id;
         this.setState({
-            selected: currentKey
+            selected: this.state.applications[key]._id
         });
-        //  console.log(currentKey);
     }
     handleDelete(){
-
         var idx = -1;
         for(let i=0; i < this.state.applications; i++){
             if(this.state.applications[i]._id == this.state.selected){
@@ -51,10 +61,33 @@ class AppLogTable extends Component{
             this.setState({applications: copy, selected: ''})
         });
     }
+    handleCompany(event){
+        this.setState({company:event.target.value});
+    }
+    handleRole(event){
+        this.setState({role:event.target.value});
+    }
+    handleStatus(event){
+        this.setState({status:event.target.value});
+    }
+    handleSubmit(event){
+        var data = {company:this.state.company,role:this.state.role,status:this.state.status};
+        //figure out how to copy applications and insert new post
+        var copy = this.state.applications.slice();
+        axios.post('/applications/create',data)
+        .then((res) => {
+            copy.push(res.data); 
+            this.setState({applications: copy})})
+        .catch(err => {console.log(err);});
+        event.preventDefault(); 
+    }   
+    // handleModify(){
 
-        //to '/applications/:ApplicationId/delete
+    // }
+
     render(){
         return(
+         <div>
             <Table onRowSelection={this.handleRowSelection} multiSelectable={this.state.multiSelectable}>
                 <TableHeader displaySelectAll={this.state.displaySelectAll} adjustForCheckbox={this.state.adjustForCheckbox}>
                     <TableRow>
@@ -74,11 +107,29 @@ class AppLogTable extends Component{
                         })
                             : null}
                    </TableBody>
-                   <TableFooter>
-                       <AppLogForm/>
-                       <FlatButton label="Delete" onClick={this.handleDelete} />
-                    </TableFooter>
             </Table>
+            <form onSubmit={this.handleSubmit} style={{textAlign:'center'}}>
+                <TextField 
+                    hintText="Company"
+                    value={this.state.company} 
+                    onChange={this.handleCompany} 
+                    style={{width:80,margin:10}}/>
+                <TextField 
+                    hintText="Role" 
+                    value={this.state.role} 
+                    onChange={this.handleRole} 
+                    style={{width:80,margin:10}}/>
+                <TextField 
+                    hintText="Status" 
+                    value={this.state.status} 
+                    onChange={this.handleStatus} 
+                    style={{width:80,margin:10}}/>
+                <br />
+                <FlatButton label="Submit" type="submit" />
+                <FlatButton label="Delete" onClick={this.handleDelete} />
+                <FlatButton label="Modify" />
+            </form>
+        </div>        
         );
     }
 }
